move trades.py to pricing_models/
