+++ elfpy/pricing_models/hyperdrive.py
@@ -6,7 +6,7 @@ from typing import TYPE_CHECKING
 
 from elfpy.pricing_models.yieldspace import YieldSpacePricingModel
 import elfpy.markets.hyperdrive as hyperdrive
-import elfpy.simulators.trades as trades
+import elfpy.pricing_models.trades as trades
 import elfpy.utils.time as time_utils
 from elfpy.agents import AgentTradeResult
 import elfpy.types as types
@@ -82,13 +82,13 @@ class HyperdrivePricingModel(YieldSpacePricingModel):
 
         .. note::
            The pool total supply is a function of the base and bond reserves, and is modified in
-           :func:`calc_lp_in_given_tokens_out 
+           :func:`calc_lp_in_given_tokens_out
            <elfpy.pricing_models.yieldspace.YieldSpacePricingModel.calc_lp_in_given_tokens_out>`,
-           :func:`calc_tokens_out_given_lp_in 
+           :func:`calc_tokens_out_given_lp_in
            <elfpy.pricing_models.yieldspace.YieldSpacePricingModel.calc_tokens_out_given_lp_in>`,
-           and :func:`calc_lp_out_given_tokens_in 
+           and :func:`calc_lp_out_given_tokens_in
            <elfpy.pricing_models.yieldspace.YieldSpacePricingModel.calc_lp_out_given_tokens_in>`.
-           
+
            It can be approximated as :math:`s \approx y + cz`.
 
         Parameters
@@ -243,13 +243,13 @@ class HyperdrivePricingModel(YieldSpacePricingModel):
 
         .. note::
            The pool total supply is a function of the base and bond reserves, and is modified in
-           :func:`calc_lp_in_given_tokens_out 
+           :func:`calc_lp_in_given_tokens_out
            <elfpy.pricing_models.yieldspace.YieldSpacePricingModel.calc_lp_in_given_tokens_out>`,
-           :func:`calc_tokens_out_given_lp_in 
+           :func:`calc_tokens_out_given_lp_in
            <elfpy.pricing_models.yieldspace.YieldSpacePricingModel.calc_tokens_out_given_lp_in>`,
-           and :func:`calc_lp_out_given_tokens_in 
+           and :func:`calc_lp_out_given_tokens_in
            <elfpy.pricing_models.yieldspace.YieldSpacePricingModel.calc_lp_out_given_tokens_in>`.
-           
+
            It can be approximated as :math:`s \approx y + cz`.
 
         Parameters
